requirement.txt
requests
pandas
sqlalchemy
python-dotenv
mysql-connector-python
psycopg2-binary